requirements.txt
numpy
Pillow
reedsolo
torch
torchvision
tqdm
